extract builtin function execution into module
